+++ src/main/java/cotato/csquiz/service/AuthService.java
@@ -5,9 +5,9 @@ import cotato.csquiz.config.jwt.RefreshToken;
 import cotato.csquiz.config.jwt.RefreshTokenRepository;
 import cotato.csquiz.config.jwt.Token;
 import cotato.csquiz.domain.dto.auth.JoinRequest;
+import cotato.csquiz.domain.dto.auth.ReissueResponse;
 import cotato.csquiz.domain.dto.email.SendSignupEmailRequest;
 import cotato.csquiz.domain.entity.Member;
-import cotato.csquiz.dto.ReissueResponse;
 import cotato.csquiz.exception.AppException;
 import cotato.csquiz.exception.ErrorCode;
 import cotato.csquiz.repository.MemberRepository;
